requirements.txt
torch>=2.0.0
numpy>=1.22.0
Pillow>=9.0.0
requests>=2.25.0 